Improve info on supported shortcodes
Some shortcodes enqueue additional CSS or Javascript, which is not
loaded via an Ajax loading method.
Added a list of known shortcodes that might have issues.
Also added an URL reset function to set the SERVER Request_URI to the
URL which made the Ajax call (and not the Ajax-URL); this increases
support for shortcodes that generate a form tag, like contact-form-7