Fix javscript security error for forms with external URL in PopUps
Some Newsletter forms have a form-action that directly goes to the
external website to submit a new member. These forms did cause a
security error after form submit, because the response code of the
external website could not be accessed by our plugin.
Now we'll just close the PopUp in case the response code is not
available and trigger a javascript event.