Improve Ajax calls to prevent security warnings by iThemes
The iThemes plugin adds a rewrite-rule that blocks all requests which
contain any URL starting with http. We send referrer information with
Ajax calls to make PopUp Referrer-Rules work with Ajax.
Solved by reversing the URLs that we send via Ajax, this way the URL
bypasses the security check by the security plugins.

URLs are only evaluated by specific rules and open no security
vulnerability if they are forged. In the worst case a PopUp is displayed
that should not have been displayed to the current user...