Improve handling of forms inside PopUps
Some users have problems where the PopUp stays open with a loading
animation after the form is submitted. Reason usually is, that the form
was submitted via Ajax and our plugin does not know the current form
state anymore and keeps the loading-animation open.

Solved by adding a timer that frequently checks if the form-target has
stopped loading (instead of simply waiting for a onLoad event - this
event is not fired when the form-target is an empty URL)